Update bank workflow tests (#130)
